go mod init, thanks to Luca
