+++ photongo/photon-go/photongo/go.mod
@@ -1,0 +1,3 @@
+module github.com/prisma/photon-go
+
+go 1.13
