+++ photongo/photon-go/photongo/prisma/user/user.go
@@ -1,6 +1,6 @@
 package user
 
-import "github.com/matthewmueller/prisma/specs/photongo/photon-go/prisma"
+import "github.com/prisma/photon-go/prisma"
 
 // Users table
 const Users = "users"
